chore: switch to btreeset for deterministic ordering (#6348)
# Description

## Problem\*

The optimisation pass from PR #6268 iterates over hash maps and hash
sets, which does not guarantee deterministic ordering. This is an issue
as it could lead to different results for the same ACIR input.

## Summary\*
Uses BTreeMap/Set when iterating.


## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.