fix(ssa): Do not mark an array from a parameter mutable (#6355)
# Description

## Problem\*

Resolves #6349 

## Summary\*

We check that we do not mark arrays mutable which comes from load
instructions as that array may be used by multiple values. This is
overridden by checking whether we are in the return block, in which case
we do mark the array as mutable. This however breaks when we no longer
inline everything. I switched the set of arrays from loads to a map of
arrays from loads to whether they come from a block parameter.

## Additional Context



## Documentation\*

Check one:
- [X] No documentation needed.
- [ ] Documentation included in this PR.
- [ ] **[For Experimental Features]** Documentation to be submitted in a
separate PR.

# PR Checklist\*

- [X] I have tested the changes locally.
- [X] I have formatted the changes with [Prettier](https://prettier.io/)
and/or `cargo fmt` on default settings.