feat(plonky2): relax the type checking convert_integer_op by allowing fields
Allows operations between a field and an integer in convert_integer_op. This
enables shift right (>>) to work, because it internally relies on
multiplication of integer and field, therefore such multiplication must be
supported by the PLONKY2 back end. I'm not sure about the other integer
operations, but we can fix them, when/if we discover any bugs.